+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.6 (22-07-2016)
+----------------
+
+Updated documentation.
+Added support for networks, network-aliases and links
+on **dockeroo:run**.
+Added new recipe **dockeroo:network**.
+
+
 0.5 (22-07-2016)
 ----------------